+++ .github/workflows/esti.yaml
@@ -1018,6 +1018,6 @@ jobs:
           verbose: true
           emoji: true
           job-summary: true
-          custom-arguments: './clients/python-wrapper/tests/system_tests -rP'
+          custom-arguments: './clients/python-wrapper/tests/integration -rP'
           click-to-expand: true
           report-title: 'Python Wrapper System Tests Report'
