+++ clients/python-wrapper/lakefs/branch_manager.py
@@ -5,6 +5,6 @@ Module implementing repository's branch management logic
 
 class BranchManager:
     """
-    Manage tags listing for a given repository
+    Manage branches listing for a given repository
     """
     # TODO: Implement
